refactor: snackbar, alert 렌더링 로직 분리
